# app/swagger.yml
swagger: "2.0"
info:
  title: Clinic Management API
  description: APIs for managing users, departments, and doctors
  version: "1.0.0"

basePath: /
schemes:
  - http
  - https

securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >
      JWT Authorization header using the Bearer scheme. Example: "Bearer {token}"

paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name, email, password]
            properties:
              name:
                type: string
              email:
                type: string
              password:
                type: string
      responses:
        201:
          description: User created successfully
        400:
          description: Email already registered

  /auth/login:
    post:
      tags:
        - Auth
      summary: Login and get JWT token
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [email, password]
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        200:
          description: Login successful
        401:
          description: Invalid credentials

  /admin/departments:
    post:
      tags:
        - Admin
      summary: Create a new department
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name]
            properties:
              name:
                type: string
      responses:
        201:
          description: Department created successfully
        400:
          description: Department already exists
    get:
      tags:
        - Admin
      summary: List all departments
      security:
        - Bearer: []
      responses:
        200:
          description: List of departments

  /admin/doctors:
    post:
      tags:
        - Admin
      summary: Onboard a new doctor
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name, email, password]
            properties:
              name:
                type: string
              email:
                type: string
              password:
                type: string
      responses:
        201:
          description: Doctor created successfully
        400:
          description: Email already exists

  /admin/doctors/assign:
    post:
      tags:
        - Admin
      summary: Assign a doctor to a department
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [doctor_id, department_id]
            properties:
              doctor_id:
                type: integer
              department_id:
                type: integer
      responses:
        201:
          description: Doctor assigned successfully
        400:
          description: Doctor already assigned